+++ package.json
@@ -3,10 +3,10 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "android": "react-native run-android",
-    "ios": "react-native run-ios",
+    "android": "eslint . && react-native run-android",
+    "ios": "eslint . && react-native run-ios",
     "lint": "eslint .",
-    "start": "react-native start",
+    "start": "eslint . && react-native start",
     "test": "jest"
   },
   "dependencies": {
