+++ src/common/firebase.js
@@ -1,17 +1,17 @@
 import messaging from "@react-native-firebase/messaging";
 import { Alert } from "react-native";
 
-export const getFcmToken = async () => {
-  const fcmToken = await messaging().getToken();
-  if (fcmToken) {
-    return fcmToken;
-  }
-};
+async function getFcmToken() {
+  await messaging().getToken();
+}
+
 export const handleNotification = (title, body) => {
   Alert.alert(title, body, [
     {
       text: "OK",
-      onPress: () => console.log("OK Pressed"),
+      onPress: () => {},
     },
   ]);
 };
+
+export { getFcmToken };
